+++ kernel/app/src/js/before.js
@@ -1,52 +1,31 @@
+import { merge } from './utils.js';
+import { Context } from './context.js';
+
 /* 覆盖 ./app/templates/config.html 配置文件中定义的字段 */
-// window.siyuan = {
-//     config: {
-//         appearance: { mode: 0, codeBlockThemeDark: "atom-one-dark", codeBlockThemeLight: "atom-one-light" },
-//         editor: {
-//             fontSize: 16,
-//             codeLineWrap: true,
-//             codeLigatures: true,
-//             codeSyntaxHighlightLineNum: true,
-//             plantUMLServePath: "https://www.plantuml.com/plantuml/svg/~1",
-//         },
-//     },
-//     languages: { copy: "复制" },
-// };
+merge(window.siyuan, {});
+merge(window.publish, {});
 
-
-/**
- * 设置编辑器字号
- * REF https://github.com/siyuan-note/siyuan/blob/fcabf93cabf0383a8b59616d66ec44e7869236cf/app/src/protyle/export/index.ts#L242-L107
- * @params {number} fontSize 字号
- * @return {number} 设置后的字号
- * @return {null} 没有找到字号
- */
-window.publish.setFontSize = (fontSize, fontFamily, codeLigatures) => {
-    let style = document.getElementById('editorFontSize');
-    if (style) {
-        const height = Math.floor(fontSize * 1.625);
-        style.innerHTML = `
-.b3-typography, .protyle-wysiwyg, .protyle-title {font-size:${fontSize}px !important}
-.b3-typography code:not(.hljs), .protyle-wysiwyg code:not(.hljs) { font-variant-ligatures: ${codeLigatures ? "normal" : "none"} }
-.li > .protyle-action {height:${height + 8}px;line-height: ${height + 8}px}
-.protyle-wysiwyg [data-node-id].li > .protyle-action ~ .h1, .protyle-wysiwyg [data-node-id].li > .protyle-action ~ .h2, .protyle-wysiwyg [data-node-id].li > .protyle-action ~ .h3, .protyle-wysiwyg [data-node-id].li > .protyle-action ~ .h4, .protyle-wysiwyg [data-node-id].li > .protyle-action ~ .h5, .protyle-wysiwyg [data-node-id].li > .protyle-action ~ .h6 {line-height:${height + 8}px;}
-.protyle-wysiwyg [data-node-id].li > .protyle-action:after {height: ${fontSize}px;width: ${fontSize}px;margin:-${fontSize / 2}px 0 0 -${fontSize / 2}px}
-.protyle-wysiwyg [data-node-id].li > .protyle-action svg {height: ${Math.max(14, fontSize - 8)}px}
-.protyle-wysiwyg [data-node-id] [spellcheck="false"] {min-height:${height}px}
-.protyle-wysiwyg .li {min-height:${height + 8}px}
-.protyle-gutters button svg {height:${height}px}
-.protyle-wysiwyg img.emoji, .b3-typography img.emoji {width:${height - 8}px}
-.protyle-wysiwyg .h1 img.emoji, .b3-typography h1 img.emoji {width:${Math.floor(fontSize * 1.75 * 1.25)}px}
-.protyle-wysiwyg .h2 img.emoji, .b3-typography h2 img.emoji {width:${Math.floor(fontSize * 1.55 * 1.25)}px}
-.protyle-wysiwyg .h3 img.emoji, .b3-typography h3 img.emoji {width:${Math.floor(fontSize * 1.38 * 1.25)}px}
-.protyle-wysiwyg .h4 img.emoji, .b3-typography h4 img.emoji {width:${Math.floor(fontSize * 1.25 * 1.25)}px}
-.protyle-wysiwyg .h5 img.emoji, .b3-typography h5 img.emoji {width:${Math.floor(fontSize * 1.13 * 1.25)}px}
-.protyle-wysiwyg .h6 img.emoji, .b3-typography h6 img.emoji {width:${Math.floor(fontSize * 1.25)}px}
-.b3-typography, .protyle-wysiwyg, .protyle-title, .protyle-title__input{font-family: ${fontFamily} "quote", "Helvetica Neue", "Luxi Sans", "DejaVu Sans", "Hiragino Sans GB", "Microsoft Yahei", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Noto Color Emoji", "Segoe UI Symbol", "Android Emoji", "EmojiSymbols" !important;}
-`;
-        return fontSize;
+/* 加载插件 */
+const context = new Context();
+for (const plugin of [...window.publish.plugin.before, ...window.publish.plugin.after]) {
+    try {
+        // 合并配置文件 custom.js
+        const module = await import(`${window.publish.plugin.path}/${plugin}/index.js`);
+        if (plugin) {
+            const plugin_obj = new module.Plugin(context);
+            window.publish.plugin.plugins.set(plugin, plugin_obj)
+        }
+    } catch (err) {
+        console.warn(err);
     }
-    return null;
+}
+
+/* 激活渲染前处理插件 */
+for (const plugin of window.publish.plugin.before) {
+    if (window.publish.plugin.plugins.has(plugin)) {
+        const plugin_obj = window.publish.plugin.plugins.get(plugin);
+        await Context.activate(plugin_obj, 'before');
+    }
 }
 
 /**
@@ -114,13 +93,6 @@ window.publish.setDocEditState = (contenteditable = false, spellcheck = false) =
 
 (() => {
     const url = new URL(window.location.href);
-
-    /* 设置字体样式 */
-    window.publish.setFontSize(
-        window.siyuan.config.editor.fontSize,
-        window.siyuan.config.editor.fontFamily,
-        window.siyuan.config.editor.codeLigatures,
-    );
 
     /* 设置题头 */
     const background = document.querySelector(".protyle-background");
@@ -198,8 +170,8 @@ window.publish.setDocEditState = (contenteditable = false, spellcheck = false) =
     /* 添加面包屑 */
     const breadcrumb = document.getElementById('breadcrumb');
     if (breadcrumb) {
-        paths = window.publish.document.path.split('/');
-        hpaths = window.publish.document.hpath.split('/');
+        const paths = window.publish.document.path.split('/');
+        const hpaths = window.publish.document.hpath.split('/');
         const depth = Math.min(paths.length, hpaths.length) - 1; // 当前文档的深度
         for (let i = 0; i <= depth; i++) {
             breadcrumb.appendChild(window.publish.createIcon(